+++ src/preload.ts
@@ -1,4 +1,20 @@
-import { InitPlugins } from "utools-helper";
 import { Config } from "./plugins/config";
 
-InitPlugins([new Config()]);
+import { IExports, IPlugin } from "./types/utools";
+
+const globalThis = window as unknown as { exports: IExports, utools: UToolsApi };
+
+const plugins: IPlugin[] = [
+    new Config(),
+];
+
+const exports: IExports = {};
+
+plugins.forEach(plugin => {
+    exports[plugin.code] = {
+        mode: plugin.mode,
+        args: plugin,
+    };
+});
+
+globalThis.exports = exports;