+++ src/types/tree.d.ts
@@ -1,0 +1,7 @@
+import { TreeNodeData } from "@arco-design/web-vue";
+
+export interface TreeNode extends TreeNodeData {
+    key: string;
+    title: string;
+    isLeaf: boolean;
+}