+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - 添加项目依赖 `utools-api-types` 与 `utools-helper` | Add dependencies `utools-api-types` and `utools-helper`.
 - 创建功能 `config` | Create a feature `config`.
 - 配置 vite 生产环境打包选项 | Config vite production build options.
+- 实现插件应用配置页面 | Implement plugin application configuration page.